@@ -8,6 +8,7 @@
     "transpile": "npm run env NODE_ENV=production && babel src --out-dir lib",
     "clean": "rimraf lib",
     "start": "cd gh-pages && webpack-dev-server --config webpack.dev.config.js --port 9090 --d --hot --inline --progress --colors",
+    "lint": "eslint src",
     "pretranspile": "npm run clean",
     "prepublish": "npm run transpile",
     "prepages-publish": "npm run env NODE_ENV=production && cd gh-pages && webpack",
@@ -29,22 +30,24 @@
   },
   "homepage": "https://github.com/effektif/react-mentions",
   "devDependencies": {
-    "babel-cli": "^6.11.4",
-    "babel-core": "^6.13.2",
+    "babel-cli": "^6.23.0",
+    "babel-core": "^6.22.1",
     "babel-loader": "^6.2.4",
-    "babel-plugin-react-transform": "^2.0.2",
     "babel-plugin-transform-class-properties": "^6.11.5",
-    "babel-plugin-transform-runtime": "^6.12.0",
-    "babel-preset-es2015": "^6.13.2",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-0": "^6.5.0",
+    "babel-plugin-transform-runtime": "^6.22.0",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-react": "^6.22.0",
+    "babel-preset-stage-0": "^6.22.0",
     "bundle-loader": "0.5.1",
-    "chai": "^3.0.0",
-    "chai-enzyme": "^0.4.1",
+    "chai": "^3.5.0",
+    "chai-enzyme": "^0.6.1",
     "chai-things": "^0.2.0",
     "css-loader": "^0.20.0",
     "envify": "1.2.0",
-    "enzyme": "^2.0.0",
+    "enzyme": "^2.7.1",
+    "eslint": "^3.16.1",
+    "eslint-config-signavio": "^1.4.2",
+    "eslint-config-signavio-test": "^1.2.1",
     "grunt": "~0.4.5",
     "grunt-contrib-copy": "^0.8.1",
     "grunt-contrib-less": "~1.0.0",
@@ -54,25 +57,25 @@
     "jsdom": "^8.0.4",
     "less": "~2.3.1",
     "less-loader": "^2.2.0",
-    "mocha": "^1.20.1",
+    "mocha": "^3.2.0",
     "mversion": "^1.10.0",
     "radium": "^0.18.1",
     "react": "^15.3.2",
     "react-addons-test-utils": "^15.3.2",
     "react-dom": "^15.3.2",
+    "react-hot-loader": "^3.0.0-beta.6",
     "react-router": "^2.6.1",
-    "react-transform-catch-errors": "^1.0.2",
-    "react-transform-hmr": "^1.0.2",
+    "recompose": "^0.22.0",
     "redbox-react": "^1.2.2",
     "rimraf": "^2.5.2",
-    "sinon": "~1.12.1",
+    "sinon": "~1.17.7",
     "sinon-chai": "^2.8.0",
-    "style-loader": "^0.12.4",
-    "webpack": "^1.12.13",
-    "webpack-dev-server": "^1.14.1"
+    "style-loader": "^0.13.1",
+    "webpack": "2.2.1",
+    "webpack-dev-server": "2.4.1"
   },
   "dependencies": {
-    "babel-runtime": "^6.11.6",
+    "babel-runtime": "^6.22.0",
     "lodash": "^4.5.1",
     "substyle": "^3.0.1"
   },
