@@ -1,6 +1,8 @@
-var React = require("react");
-var ReactDOM = require("react-dom");
-var Application = require("./views/Application");
-require("./less/react-mentions.less");
+import React from 'react'
+import ReactDOM from 'react-dom'
 
-ReactDOM.render(React.createElement(Application, null), document.getElementById('app'));
+import './less/react-mentions.less'
+
+import Application from './views'
+
+ReactDOM.render(<Application />, document.getElementById('app'))
